--- src/vercel.json
@@ -1,22 +1,0 @@
-{
-    "name": "backendNode",
-    "version": 1,
-    "builds": [
-      {
-        "src": "index.js",
-        "use": "@vercel/node"
-      }
-    ],
-    "routes": [
-      {
-        "src": "/(.*)",
-        "dest": "/index.js"
-      }
-    ],
-    "env": {
-      "DB_USER": "@backendnode-db-user",
-      "DB_PASSWORD": "@backendnode-db-password",
-      "DB_HOST": "@backendnode-db-host",
-      "DB_NAME": "@backendnode-db-name"
-    }
-  }
